Add files via upload
I fixed the side patterns after it kept getting cut off and fixed the position as well. I also added information to the separate portfolio pages. I need help making the panels on the right side of the portfolio page to be at that position that the viewer is already at on the page instead of being all the way at the top.